FileUpload guts in, need status windows and what not
Tested with 6gb file, handled it.

+++ master/FileUpload/FileUpload/Areas/Orders/File.cs
@@ -2,7 +2,7 @@
 {
 	public class File
 	{
-		public int FileId { get; set; }
+		public long FileId { get; set; }
 		public string Name { get; set; } = string.Empty;
 		public string FilePath { get; set; } = string.Empty;
         public string ContentType { get; set; } = string.Empty;
@@ -10,7 +10,7 @@
 		public long Length { get; set; } = 0;
 
 		public long Written { get; set; } = 0;
-        public int OrderId { get; set; }
+        public long OrderId { get; set; }
 		public Order Order { get; set; }
 		public File(string Name, string FilePath) { 
 			this.Name = Name;
